Struct crypto_vault::Vault

pub struct Vault<T: Decodable + Encodable + Debug> {
    pub objects: Vec<T>,
    // some fields omitted
}

Stores the decrypted data. May have the master key and salt for encryption.

Fields

objects

The decrypted data

Methods

impl<T: Decodable + Encodable + Debug> Vault<T>

fn set_password(&mut self, password: &str)

Sets the password. Overrides previous password.

fn with_password(self, password: &str) -> Vault<T>

Sets the password. Overrides previous password.

Chainable version of set_password

fn encrypt(&self) -> VResult<RawVault>

Encrypts the store into a RawVault.

Returns a VaultError::NoPasswordSpecifiedError if there is no password set on the store.

fn new() -> Vault<T>

Constructor function